requirements.txt
tensorflow-gpu==2.10.0
tensorflow_model_optimization

opencv-Python
matplotlib
seaborn
scikit-learn
tqdm